Add Descriptions
For each English word as a key, the value is a list of 4 elements. The first and third elements are a list of all correct translations for each sense, respectively. The second element is an example sentence for the first sense. The fourth element is an example sentence for the second sense.